+++ src/gui/render.cpp
@@ -648,7 +648,7 @@ static void IncreaseFrameSkip(bool pressed) {
 	if (!pressed)
 		return;
 	if (render.frameskip.max<10) render.frameskip.max++;
-	LOG_MSG("Frame Skip at %d",render.frameskip.max);
+	LOG_MSG("Frame Skip at %d",(int)render.frameskip.max);
 	GFX_SetTitle(-1,render.frameskip.max,-1,false);
 }
 
@@ -656,7 +656,7 @@ static void DecreaseFrameSkip(bool pressed) {
 	if (!pressed)
 		return;
 	if (render.frameskip.max>0) render.frameskip.max--;
-	LOG_MSG("Frame Skip at %d",render.frameskip.max);
+	LOG_MSG("Frame Skip at %d",(int)render.frameskip.max);
 	GFX_SetTitle(-1,render.frameskip.max,-1,false);
 }
 /* Disabled as I don't want to waste a keybind for that. Might be used in the future (Qbix)
