+++ src/hardware/serialport/softmodem.h
@@ -91,7 +91,7 @@ public:
 			static Bits lcount=0;
 			if (lcount<1000) {
 				lcount++;
-				LOG_MSG("MODEM: FIFO Overflow! (adds len %u)",_len);
+				LOG_MSG("MODEM: FIFO Overflow! (adds len %u)",(int)_len);
 			}
 			return;
 		}
@@ -125,7 +125,7 @@ public:
 			static Bits lcount=0;
 			if (lcount<1000) {
 				lcount++;
-				LOG_MSG("MODEM: FIFO UNDERFLOW! (gets len %d)",_len);
+				LOG_MSG("MODEM: FIFO UNDERFLOW! (gets len %d)",(int)_len);
 			}
 			return;
 		}
